FCL-163 | split out details views
