+++ judgments/tests/test_document_resolver_engine.py
@@ -6,23 +6,23 @@ from judgments.resolvers.document_resolver_engine import DocumentResolverEngine
 
 
 class TestDocumentResolverEngine(TestCase):
-    @patch("judgments.resolvers.document_resolver_engine.get_best_pdf")
-    @patch("judgments.resolvers.document_resolver_engine.get_generated_pdf")
+    @patch("judgments.resolvers.document_resolver_engine.best_pdf")
+    @patch("judgments.resolvers.document_resolver_engine.generated_pdf")
     @patch("judgments.resolvers.document_resolver_engine.detail_xml")
-    @patch("judgments.resolvers.document_resolver_engine.detail")
+    @patch("judgments.resolvers.document_resolver_engine.detail_html")
     def test_resolver_engine_with_fileformats(
         self,
-        mock_detail,
+        mock_detail_html,
         mock_detail_xml,
-        mock_get_generated_pdf,
-        mock_get_best_pdf,
+        mock_generated_pdf,
+        mock_best_pdf,
     ):
         document_uri = "ewhc/comm/2024/253"
         test_params = [
-            ("data.pdf", mock_get_best_pdf),
-            ("generated.pdf", mock_get_generated_pdf),
+            ("data.pdf", mock_best_pdf),
+            ("generated.pdf", mock_generated_pdf),
             ("data.xml", mock_detail_xml),
-            ("data.html", mock_detail),
+            ("data.html", mock_detail_html),
         ]
         for file_format, view in test_params:
             with self.subTest(filename=file_format, view=view):
