+++ judgments/utils/__init__.py
@@ -1,3 +1,4 @@
+from .judgment_utils import get_published_document_by_uri
 from .search_utils import (
     ALL_COURT_CODES,
     get_minimum_valid_year,
@@ -34,6 +35,7 @@ __all__ = [
     "clamp",
     "formatted_document_uri",
     "get_document_by_uri",
+    "get_published_document_by_uri",
     "get_minimum_valid_year",
     "get_press_summaries_for_document_uri",
     "has_filters",
