+++ judgments/views/detail/detail_html.py
@@ -1,24 +1,18 @@
 import logging
 import os
 import urllib
-from typing import Any, Union
+from typing import Any
 
-import requests
-from caselawclient.errors import DocumentNotFoundError, MarklogicNotPermittedError
-from caselawclient.models.documents import Document, DocumentURIString
-from django.conf import settings
-from django.http import Http404, HttpResponse, HttpResponseRedirect
-from django.shortcuts import redirect
+from caselawclient.models.documents import DocumentURIString
+from django.http import Http404, HttpResponseRedirect
 from django.template.defaultfilters import filesizeformat
 from django.template.response import TemplateResponse
 from django.urls import reverse
-from django.views.generic import TemplateView
-from django_weasyprint import WeasyTemplateResponseMixin
 
 from judgments.forms import AdvancedSearchForm
 from judgments.models.document_pdf import DocumentPdf
 from judgments.utils import (
-    get_document_by_uri,
+    get_published_document_by_uri,
     linked_doc_title,
     linked_doc_url,
     preprocess_query,
@@ -37,69 +31,7 @@ if os.environ.get("SHOW_WEASYPRINT_LOGS") != "True":
     logging.getLogger("weasyprint").handlers = []
 
 
-def get_published_document_by_uri(document_uri: Union[str, None], cache_if_not_found: bool = False) -> Document:
-    if not document_uri:
-        raise Http404("Missing document uri")
-    try:
-        document = get_document_by_uri(document_uri, cache_if_not_found=cache_if_not_found)
-        if not document:
-            raise Http404(f"Document {document_uri} was not found")
-    except DocumentNotFoundError:
-        raise Http404(f"Document {document_uri} was not found")
-    except MarklogicNotPermittedError:
-        raise Http404(f"Document {document_uri} is not available")
-
-    if not document.is_published:
-        raise Http404(f"Document {document_uri} is not available")
-    return document
-
-
-class PdfDetailView(WeasyTemplateResponseMixin, TemplateView):
-    template_name = "pdf/document.html"
-    pdf_stylesheets = [os.path.join(settings.STATIC_ROOT, "css", "documentpdf.css")]
-    pdf_attachment = True
-
-    def get_context_data(self, document_uri=None, **kwargs) -> dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-
-        document = get_published_document_by_uri(document_uri)
-
-        self.pdf_filename = f"{document.uri}.pdf"
-
-        context["document"] = document.content_as_html(MOST_RECENT_VERSION)
-
-        return context
-
-
-def get_generated_pdf(request, document_uri):
-    return PdfDetailView.as_view()(request, document_uri=document_uri)
-
-
-def get_best_pdf(request, document_uri):
-    """
-    Response for the legacy data.pdf endpoint, used by data reusers
-
-    If there's a DOCX-derived PDF in the S3 bucket, return that.
-    Otherwise fall back and redirect to the weasyprint version."""
-    pdf = DocumentPdf(document_uri)
-    response = requests.get(pdf.generate_uri())
-    logging.debug("Response %s", response.status_code)
-    if response.status_code == 200:
-        return HttpResponse(response.content, content_type="application/pdf")
-
-    if response.status_code != 404:
-        logging.warn(f"Unexpected {response.status_code} error on {document_uri} whilst trying to get_best_pdf")
-    # fall back to weasy_pdf
-
-    return redirect(
-        reverse(
-            "detail",
-            kwargs={"document_uri": document_uri, "file_format": "generated.pdf"},
-        )
-    )
-
-
-def detail(request, document_uri):
+def detail_html(request, document_uri):
     query = request.GET.get("query")
     document = get_published_document_by_uri(document_uri)
     pdf = DocumentPdf(document_uri)
@@ -166,13 +98,3 @@ def detail(request, document_uri):
     context["pdf_uri"] = pdf.uri  # TODO: Remove this from context
 
     return TemplateResponse(request, "judgment/detail.html", context=context)
-
-
-def detail_xml(_request, document_uri):
-    document = get_published_document_by_uri(document_uri)
-
-    document_xml = document.content_as_xml
-
-    response = HttpResponse(document_xml, content_type="application/xml")
-    response["Content-Disposition"] = f"attachment; filename={document.uri}.xml"
-    return response
